+++ frontend/src/components/list_editor.tsx
@@ -5,6 +5,7 @@ import { getThings } from '../api/things';
 import { Pages } from './pages';
 import { AuthContext } from '../context/auth';
 import { AxiosContext } from '../context/axios';
+import { SharingStateComponent } from './shared/sharing_state';
 
 export type ListEditorData = {
   name: string;
@@ -14,15 +15,27 @@ export type ListEditorData = {
 
 type ListEditorProps = {
   children?: ReactNode;
-  list: ListEditorData;
+  list?: ListEditorData;
   onChange: (list: ListEditorData) => void;
+  defaultSharingState?: SharingState;
 };
 
 const thingsPerPage = 10;
 
-export const ListEditor = ({ children, list, onChange }: ListEditorProps) => {
+export const ListEditor = ({
+  children,
+  list,
+  onChange,
+  defaultSharingState = 'private',
+}: ListEditorProps) => {
   const authCtx = useContext(AuthContext);
   const axiosInstance = useContext(AxiosContext);
+  const [name, setName] = useState('');
+  const [selectedThingIDs, setSelectedThingIDs] = useState<string[]>([]);
+  const [sharingStateAndDefault, setSharingStateAndDefault] = useState<[SharingState, boolean]>([
+    'private',
+    true,
+  ]);
   const [currentPage, setCurrentPage] = useState(0);
   const [selectedPage, setSelectedPage] = useState(0);
   const [textFilter, setTextFilter] = useState('');
@@ -31,31 +44,46 @@ export const ListEditor = ({ children, list, onChange }: ListEditorProps) => {
     undefined
   );
 
+  const [sharingState, usingDefaultSharingState] = sharingStateAndDefault;
+
+  useEffect(() => {
+    if (!list) {
+      return;
+    }
+    setName(list.name);
+    setSelectedThingIDs(list.selectedThingIDs);
+    setSharingStateAndDefault([list.sharingState, false]);
+  }, [list]);
+
+  useEffect(() => {
+    if (list) {
+      // If we have a loaded list, don't override its sharing state
+      return;
+    }
+    if (defaultSharingState) {
+      setSharingStateAndDefault([defaultSharingState, true]);
+    }
+  }, [defaultSharingState, list]);
+
+  useEffect(() => {
+    const data = {
+      name,
+      selectedThingIDs,
+      sharingState,
+    };
+    onChange(data);
+  }, [name, selectedThingIDs, sharingState, onChange]);
+
   const onThingSelect = (thingID: string, isChecked: boolean) => {
-    const selectedThingIDs = list.selectedThingIDs;
     if (isChecked) {
-      onChange({
-        ...list,
-        selectedThingIDs: [...selectedThingIDs, thingID],
-      });
+      setSelectedThingIDs([...selectedThingIDs, thingID]);
     } else {
-      const index = selectedThingIDs.indexOf(thingID);
-      if (index > -1) {
-        const updatedSelectedThingIDs = [...selectedThingIDs];
-        updatedSelectedThingIDs.splice(index, 1);
-        onChange({
-          ...list,
-          selectedThingIDs: updatedSelectedThingIDs,
-        });
-      }
+      setSelectedThingIDs(selectedThingIDs.filter((id) => id !== thingID));
     }
   };
 
   const onNameChange = (value: string) => {
-    onChange({
-      ...list,
-      name: value,
-    });
+    setName(value);
   };
 
   useEffect(() => {
@@ -78,15 +106,15 @@ export const ListEditor = ({ children, list, onChange }: ListEditorProps) => {
 
   const selectedThingIDsPage = useMemo(() => {
     const start = selectedPage * thingsPerPage;
-    return list.selectedThingIDs.slice(start, start + thingsPerPage);
-  }, [list.selectedThingIDs, selectedPage]);
+    return selectedThingIDs.slice(start, start + thingsPerPage);
+  }, [selectedThingIDs, selectedPage]);
 
   const onTextFilterChange = (filter: string) => {
     setCurrentPage(0);
     setTextFilter(filter);
   };
 
-  const selectedTotalPages = Math.ceil(list.selectedThingIDs.length / thingsPerPage);
+  const selectedTotalPages = Math.ceil(selectedThingIDs.length / thingsPerPage);
 
   return (
     <div>
@@ -99,13 +127,24 @@ export const ListEditor = ({ children, list, onChange }: ListEditorProps) => {
             type="text"
             id="name"
             name="name"
-            value={list.name}
+            value={name}
             onChange={(e) => onNameChange(e.target.value)}
             className="mt-1 p-2 text-display border border-gray-300 rounded-sm"
           />
         </div>
         <div className="flex flex-col justify-end">{children}</div>
       </div>
+      {!list && (
+        <div className="bg-neutral-900 border border-neutral-700 rounded-sm p-3 my-2 w-100">
+          <p className="text-display text-sm flex">
+            Sharing: <SharingStateComponent state={sharingState} />
+            {usingDefaultSharingState && (
+              <span className="text-secondary text-sm">(from your profile default)</span>
+            )}
+          </p>
+        </div>
+      )}
+
       <div className="flex gap-4 items-stretch">
         <div className="flex-1 flex flex-col justify-between">
           <div>
@@ -115,7 +154,7 @@ export const ListEditor = ({ children, list, onChange }: ListEditorProps) => {
                 <SelectableThing
                   key={thing.id}
                   thing={thing}
-                  selected={list.selectedThingIDs.includes(thing.id)}
+                  selected={selectedThingIDs.includes(thing.id)}
                   onSelect={onThingSelect}
                 />
               ))}
@@ -142,7 +181,7 @@ export const ListEditor = ({ children, list, onChange }: ListEditorProps) => {
         <div className="flex-1 flex flex-col justify-between">
           <div>
             <h3 className="text-primary text-sm font-medium mb-2">
-              Selected ({list.selectedThingIDs.length})
+              Selected ({selectedThingIDs.length})
             </h3>
             <div className="flex flex-wrap gap-4 content-start">
               {selectedThingIDsPage.map((thingId) => (