+++ frontend/src/routes/lists/create.tsx
@@ -1,22 +1,20 @@
 import { useContext, useState } from 'react';
 import { useNavigate } from 'react-router';
 import { AxiosContext } from '../../context/axios';
+import { AuthContext } from '../../context/auth';
 import { ListEditor, ListEditorData } from '../../components/list_editor';
 import { createList } from '../../api/lists';
 import { PrimaryButton } from '../../components/shared';
 
 export const CreateList = () => {
   const axiosInstance = useContext(AxiosContext);
+  const authContext = useContext(AuthContext);
   const navigate = useNavigate();
 
-  const [editedData, setEditedData] = useState<ListEditorData>({
-    name: '',
-    selectedThingIDs: [],
-    sharingState: 'private',
-  });
+  const [editedData, setEditedData] = useState<null | ListEditorData>(null);
 
   const create = async () => {
-    if (!axiosInstance) {
+    if (!axiosInstance || !editedData) {
       return;
     }
     const list = await createList(axiosInstance, {
@@ -29,7 +27,10 @@ export const CreateList = () => {
   };
 
   return (
-    <ListEditor list={editedData} onChange={setEditedData}>
+    <ListEditor
+      onChange={setEditedData}
+      defaultSharingState={authContext.profile?.defaultSharingState}
+    >
       <PrimaryButton onClick={create}>Create</PrimaryButton>
     </ListEditor>
   );