+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- add defaultSharingState to profile, creates things and lists with that sharing setting
 - development: a nix-based development environment is available
   with the `dev` flake output.
 - backend: add OIDC auth mechanism